+++ dev/include/ofdm_core.h
@@ -4,10 +4,10 @@
 
 bool is_guard(int k, SharedData &sd);
 
-vector<complex<double>> insert_pilots(const vector<complex<double>>& symbols, SharedData& sd);
-vector<complex<double>> ofdm_modulator(const vector<complex<double>>& freq_symbols, SharedData& sd);
-vector<complex<double>> discard_cp(vector<complex<double>> signal, SharedData &sd);
+vector<complex<float>> insert_pilots(const vector<complex<float>>& symbols, SharedData& sd);
+vector<complex<float>> ofdm_modulator(const vector<complex<float>>& freq_symbols, SharedData& sd);
+vector<complex<float>> discard_cp(vector<complex<float>> signal, SharedData &sd);
 void update_pilots(SharedData& sd);
-vector<complex<double>> ofdm_equalize(const vector<complex<double>>& signal, SharedData &sd);
+vector<complex<float>> ofdm_equalize(const vector<complex<float>>& signal, SharedData &sd);
 
 #endif
